Floor to udp

+++ src/main/Instruction.java
@@ -1,4 +1,7 @@
 package main;
+
+import FloorSubsystem.FloorDirection;
+
 /**
  * This class will be the datatype used to transfer and handle the input file instructions
  * I decided on a custom datatype to make using a queue simpler and it allows for expansion when we reach the fault iteration.
@@ -37,9 +40,9 @@ public class Instruction {
 			System.out.println("Invalid Time\n");
 			return -1;
 		}
-		int hours = Integer.parseInt(times[0]);
-		int mins = Integer.parseInt(times[1]);
-		int seconds = Integer.parseInt(times[2]);
+		double hours = Double.parseDouble(times[0]);
+		double mins = Double.parseDouble(times[1]);
+		double seconds = Double.parseDouble(times[2]);
 		return (hours * 3600) + (mins * 60) + seconds;
 	}
